QuickHack 2020 - Iconnect
IConnect.html

Team Name: AOET DUO
Team Member: Ashley Ooi, Tan Eng Teck

Overview:

Due to the Covid-19 outbreak, education can’t be delivered through face-to-face meetings. It is a challenge for students to study online. There are many students struggling in heavy workshops and online classes. Therefore, students should have a confession platform to meet new friends and help each other. 

Description: 
IConnect is an online forum which connects Monash past, present and future students. It allows students to share their ideas and opinions. Students are able to interact with each student with an on campus feeling through this forum. Besides that, students can be alerted with the announcements from the institution and concerns of Covid-19. 

Objectives:
1. To connect with Monash past, present and future students.
2. To raise awareness of Covid-19. 

Approach:

1. Create a Forum to interact with each other.
2. COVID-19 concern updated daily.
3. Group discussion with small topics. 
4. Create a small chat box for users for pm purposes. 

Impact and solution
1. Online platform such as forum
   1. Students can interact with each other and keep updated with school announcements. 
   2. Students can make posts related to events or activities.
   3. Strong connection between students in this Covid-19 pandemic
   4. Students are able to seek help/advice from seniors/peers.
   5. To help build a sense of community online
   6. Students are able to learn from one another


2. Covid-19 concern
   1. Keep updated with cases in Malaysia
   2. Able to follow up the announcements from Monash University
   3. Provide counselling for students who are affected by this pandemic.
   4. Provide reminders on the precautionary measures


3. Team Available
   1. For students who are in the same courses to discuss
   2. Group discussion with minimal instructor involvement

4. Chat box
   1. Used for asking information between students.
   2. Direct message from students to the representative of a particular group.